Extract sfuture library.
https://github.com/sgkim126/sfuture.ts
