+++ package.json
@@ -34,6 +34,7 @@
     "mongodb": "^2.0.30",
     "mpromise": "^0.5.4",
     "node-uuid": "^1.4.3",
+    "sfuture": "0.0.4",
     "typescript": "^1.5.0-alpha"
   },
   "devDependencies": {
